--- lib/src/sample_feature/sample_item.dart
@@ -1,6 +1,0 @@
-/// A placeholder class that represents an entity or model.
-class SampleItem {
-  const SampleItem(this.id);
-
-  final int id;
-}
