--- lib/src/sample_feature/sample_item_details_view.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Displays detailed information about a SampleItem.
-class SampleItemDetailsView extends StatelessWidget {
-  const SampleItemDetailsView({super.key});
-
-  static const routeName = '/sample_item';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Item Details'),
-      ),
-      body: const Center(
-        child: Text('More Information Here'),
-      ),
-    );
-  }
-}
