+++ lib/src/settings/settings_view.dart
@@ -17,7 +17,7 @@ class SettingsView extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Some shit'),
+        title: const Text('Choose theme'),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16),
